backend/requirements.txt
fastapi
uvicorn
solana
pydantic
base58
structlog
loguru
python-multipart
pillow